bug fixes, reformatting
move readycheck to separate file
deprecate moviepoll
add better error handling for api calls
