+++ players.json
@@ -5,7 +5,7 @@
     },
     "287781427733331968": {
         "name": "cruel efficiency",
-        "id": "488684811"
+        "id": "74675113"
     },
     "124311920814587904": {
         "name": "shigus",
@@ -94,5 +94,13 @@
     "105819872780419072" : {
         "name": "Mercury Sulfide",
         "id": "82444103"
+    },
+    "514663058493276162" : {
+        "name": "tlace",
+        "id": "237910815"
+    },
+    "207539633117069313" : {
+        "name": "satman",
+        "id": "91504727"
     }
 }
